Rewrite README in Markdown
The original motivation for writing the README in free-form format was
that I found Markdown's syntax for links too distracting when reading
the source.  However, I think the pretty formatting that it wins you
in the web interface is worth it.

All that I really need is footnotes, but Markdown doesn't offer those,
and the markup languages supported by GitHub that do have too busy
syntaxes for them, I think.